# go_resource_abstractor/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Resource Abstractor API
  version: v1
  description: |
    REST API in front of MongoDB for reading and writing Oakestra's resource
    state: candidate resources (workers at cluster level, clusters at root
    level), applications, jobs and their per-worker instances, webhook
    registrations, and dynamically registered custom resource types.

    The same service runs at both the root (`root_resource_abstractor`, port
    11011) and cluster (`cluster_resource_abstractor`, port 11012) level,
    differentiated purely by env vars.

    ## Schemaless documents

    Every document collection is a pass-through to MongoDB: the request bodies
    below list the fields the service itself reads or validates, but unknown
    fields are accepted and stored verbatim, and returned again on reads. That
    is why every document schema sets `additionalProperties: true` - treat the
    listed properties as the fields consumers can rely on, not as an exhaustive
    list of what a stored document contains.

    ## Trailing slashes

    Every path is also reachable with a single trailing slash (`/api/v1/jobs/`
    resolves identically to `/api/v1/jobs`), so clients written against the
    original Flask service - which set `strict_slashes = False` - keep working
    unchanged.
servers:
  - url: http://localhost:11011
    description: Root orchestrator (root_resource_abstractor)
  - url: http://localhost:11012
...
paths:
...
  /api/v1/resources:
    get:
      operationId: listResources
      tags: [Resources]
      summary: List candidate resources
      description: |
        Returns candidates with a computed `active` flag (true when
        `last_modified_timestamp` is within the last 30 seconds) and a
        projection limited to the canonical resource fields, extendable via
        `resources`.
      parameters:
        - $ref: '#/components/parameters/ResourceActive'
        - $ref: '#/components/parameters/ResourceJobID'
        - $ref: '#/components/parameters/ResourceCandidateName'
        - $ref: '#/components/parameters/ResourceIP'
        - $ref: '#/components/parameters/ResourceProjection'
      responses:
        '200':
          $ref: '#/components/responses/ResourceList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidInput'
    post:
      operationId: createResource
      tags: [Resources]
      summary: Create a candidate resource
      requestBody:
        $ref: '#/components/requestBodies/Resource'
      responses:
        '201':
          $ref: '#/components/responses/Resource'
        '422':
          $ref: '#/components/responses/InvalidInput'
    put:
      operationId: upsertResource
      tags: [Resources]
      summary: Create or update a candidate resource by name
      description: |
        Updates the candidate whose `candidate_name` matches the body, or
        creates a new one if no candidate carries that name. This is the call
        a cluster manager makes on every resource-aggregation tick.
      requestBody:
        $ref: '#/components/requestBodies/Resource'
      responses:
        '200':
          $ref: '#/components/responses/Resource'
        '422':
          $ref: '#/components/responses/InvalidInput'
  /api/v1/resources/{id}:
    parameters:
      - $ref: '#/components/parameters/ObjectID'
    get:
      operationId: getResource
      tags: [Resources]
      summary: Fetch one candidate resource
      responses:
        '200':
          $ref: '#/components/responses/Resource'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchResource
      tags: [Resources]
      summary: Report resource usage for a candidate
      description: |
        Persists an aggregated usage report: `last_modified` /
        `last_modified_timestamp` are refreshed server-side, the reported
        `cpu_percent` and `memory_percent` are appended to `cpu_history` and
        `memory_history` (each capped at the most recent 100 samples), and the
        rest of the body is stored as-is.

        Any `cpu_history` / `memory_history` sent by the client is ignored, so
        echoing a previously read document back does not duplicate history.

        Note that an `id` that is not a valid ObjectID is reported as 404 here,
        not 400 - matching the original Flask service.
      requestBody:
        $ref: '#/components/requestBodies/Resource'
      responses:
        '200':
          $ref: '#/components/responses/Resource'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidInput'
    delete:
      operationId: deleteResource
      tags: [Resources]
      summary: Delete a candidate resource
      responses:
        '204':
          description: Deleted. Also returned when no candidate carried that id.
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      summary: List jobs
      parameters:
        - $ref: '#/components/parameters/JobApplicationID'
        - $ref: '#/components/parameters/JobName'
      responses:
        '200':
          description: Matching jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
    post:
      operationId: createJob
      tags: [Jobs]
      summary: Create a job
      requestBody:
        $ref: '#/components/requestBodies/Job'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
    put:
      operationId: upsertJob
      tags: [Jobs]
      summary: Create or update a job by name
      description: |
        Updates the job whose `job_name` matches the body, or creates a new one
        if no job carries that name.
      requestBody:
        $ref: '#/components/requestBodies/Job'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/JobID'
    get:
      operationId: getJob
      tags: [Jobs]
      summary: Fetch one job
      parameters:
        - $ref: '#/components/parameters/JobInstanceNumber'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidInput'
    patch:
      operationId: patchJob
      tags: [Jobs]
      summary: Update a job
      requestBody:
        $ref: '#/components/requestBodies/Job'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteJob
      tags: [Jobs]
      summary: Delete a job
      responses:
        '200':
          description: The job as it was immediately before deletion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/jobs/{job_id}/{instance_id}:
    parameters:
      - $ref: '#/components/parameters/JobID'
      - $ref: '#/components/parameters/InstanceID'
    get:
      operationId: getJobInstance
      tags: [Jobs]
      summary: Fetch one job instance
      description: |
        Returns the job, with `instance_list` filtered down to just the
        matching instance - not a bare instance object, the same shape
        `put`/`patch`/`delete` on this path return.
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: appendJobInstance
      tags: [Jobs]
      summary: Append a job instance
      description: |
        Appends the last element of the body's `instance_list` to the job,
        stamped with `instance_id` as its `instance_number`. Appending an
        instance number the job already carries is rejected with 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobInstanceAppend'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
    patch:
      operationId: patchJobInstance
      tags: [Jobs]
      summary: Update a job instance
      description: |
        Updates the instance's status and resource-usage fields in place, and
        appends the reported `cpu` / `memory` values to that instance's
        history arrays (each capped at the most recent 100 samples).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobInstance'
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteJobInstance
      tags: [Jobs]
      summary: Delete a job instance
      responses:
        '200':
          $ref: '#/components/responses/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    ObjectID:
      name: id
      in: path
      required: true
      description: Hex-encoded MongoDB ObjectID.
      schema:
        $ref: '#/components/schemas/ObjectID'
    JobID:
      name: job_id
      in: path
      required: true
      description: Hex-encoded MongoDB ObjectID of the job.
      schema:
        $ref: '#/components/schemas/ObjectID'
    InstanceID:
      name: instance_id
      in: path
      required: true
      description: The instance's `instance_number` within its job.
      schema:
        type: integer
    ResourceActive:
      name: active
      in: query
      required: false
      description: |
        When truthy, only return candidates that reported within the freshness
        window (30 seconds).

        Typed as a string rather than a boolean because the original Flask
        service parsed it with marshmallow, which accepts `1`, `t`, `true`,
        `on`, `y`, `yes` and their false counterparts (`0`, `f`, `false`,
        `off`, `n`, `no`), case-insensitively. Anything else is a 422.
      schema:
        type: string
    ResourceJobID:
      name: job_id
      in: query
      required: false
      description: |
        Return only the candidate the given job is placed on. Resolved through
        the job's `candidate` field, so an unknown job or one not yet placed
        is a 404.
      schema:
        $ref: '#/components/schemas/ObjectID'
    ResourceCandidateName:
      name: candidate_name
      in: query
      required: false
      schema:
        type: string
    ResourceIP:
      name: ip
      in: query
      required: false
      schema:
        type: string
    ResourceProjection:
      name: resources
      in: query
      required: false
      description: |
        Extra document fields to include beyond the canonical projection, as a
        comma-separated list.
      style: form
      explode: false
      schema:
        type: array
        items:
          type: string
      example: [gpu_temp, gpu_drivers]
...
    JobApplicationID:
      name: applicationID
      in: query
      required: false
      schema:
        type: string
    JobName:
      name: job_name
      in: query
      required: false
      schema:
        type: string
    JobInstanceNumber:
      name: instance_number
      in: query
      required: false
      description: |
        Narrow the result to jobs carrying an instance with this number.

        Typed as a numeric string rather than an integer so that a
        non-numeric value keeps returning the 422 `Invalid input` body the
        original Flask service returned, rather than a plain 400.
      schema:
        type: string
        pattern: '^-?[0-9]+$'
...
  requestBodies:
    Resource:
      required: false
      description: An absent or empty body is accepted and treated as `{}`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Resource'
...
    Job:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Job'
...
  responses:
    Resource:
      description: The stored candidate resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Resource'
    ResourceList:
      description: Matching candidate resources.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Resource'
...
    Job:
      description: The stored job, including its full instance list.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Job'
...
    BadRequest:
      description: Malformed request - an unparseable body, or an id that is not a valid ObjectID.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    NotFound:
      description: No such document.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    InvalidInput:
      description: The request body or query parameters failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
    ObjectID:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
      example: 65f1c0d2a1b2c3d4e5f60718
    Message:
      type: object
      description: A bare status or error message.
      required: [message]
      properties:
        message:
          type: string
      example:
        message: Not Found
    ValidationError:
      type: object
      description: |
        A validation failure, carrying per-field messages under `details`. The
        shape of `details` follows the failing location: `{"body": {...}}` for
        request bodies and `{"query": {...}}` for query parameters.
      required: [message, details]
      properties:
        message:
          type: string
          example: Invalid input
        details:
          type: object
          additionalProperties: true
      example:
        message: Invalid input
        details:
          query:
            active: ['active must be a valid boolean']
...
    Resource:
      type: object
      description: |
        A candidate resource: a worker node at cluster level, a whole cluster
        at root level. Unknown fields are stored and returned verbatim.
      additionalProperties: true
      properties:
        _id:
          x-go-name: ID
          allOf:
            - $ref: '#/components/schemas/ObjectID'
        candidate_name:
          type: string
        candidate_location:
          type: string
        ip:
          type: string
        port:
          type: string
        architecture:
          type: string
        active:
          type: boolean
          readOnly: true
          description: |
            True when `last_modified_timestamp` falls within the last 30
            seconds. Computed per request rather than persisted, and outside
            the canonical projection - so it is only present on
            `GET /api/v1/resources` and only when asked for explicitly, with
            `?resources=active`.
        active_nodes:
          type: integer
        memory:
          type: integer
        vcpus:
          type: integer
        vgpus:
          type: integer
        cpu_percent:
          type: number
          format: double
        memory_percent:
          type: number
          format: double
        gpu_percent:
          type: integer
        aggregation_per_architecture:
          type: object
          additionalProperties: true
        virtualization:
          type: array
          nullable: true
          items:
            type: string
        supported_addons:
          type: array
          nullable: true
          items:
            type: string
        csi_drivers:
          type: array
          nullable: true
          items: {}
        cpu_history:
          type: array
          readOnly: true
          description: Most recent 100 cpu samples, appended server-side.
          items:
            $ref: '#/components/schemas/HistorySample'
        memory_history:
          type: array
          readOnly: true
          description: Most recent 100 memory samples, appended server-side.
          items:
            $ref: '#/components/schemas/HistorySample'
        last_modified_timestamp:
          type: number
          format: double
          description: Unix seconds, refreshed server-side on every usage report.
    HistorySample:
      type: object
      properties:
        value:
          type: number
          format: double
          nullable: true
        timestamp:
          type: number
          format: double
    JobHistorySample:
      type: object
      description: |
        A job instance's cpu/memory history sample. Unlike Resource's
        HistorySample, timestamp is an ISO-8601 string
        (datetime.now(timezone.utc).isoformat()), not a Unix-seconds number -
        the two aren't interchangeable.
      properties:
        value:
          type: number
          format: double
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    Job:
      type: object
      description: |
        A deployed microservice and its per-worker instances. Jobs carry the
        whole SLA the user submitted, so unknown fields are the norm here -
        they are stored and returned verbatim.
      additionalProperties: true
      properties:
        _id:
          x-go-name: ID
          allOf:
            - $ref: '#/components/schemas/ObjectID'
        job_name:
          type: string
          description: Unique per deployment; the key `PUT /api/v1/jobs` upserts on.
        applicationID:
          type: string
        candidate:
          type: string
          description: Id of the candidate resource this job is placed on.
        instance_list:
          type: array
          items:
            $ref: '#/components/schemas/JobInstance'
    JobInstance:
      type: object
      description: One running copy of a job on one worker.
      additionalProperties: true
      properties:
        instance_number:
          type: integer
        worker_id:
          type: string
        status:
          type: string
        status_detail:
          type: string
        host_ip:
          type: string
        host_port:
          type: integer
        publicip:
          type: string
        cpu_percent:
          type: number
          format: double
        memory_percent:
          type: number
          format: double
        disk:
          type: number
          format: double
        logs:
          type: string
        last_modified_timestamp:
          type: number
          format: double
          description: |
            Unix seconds. Unlike cpu_history/memory_history's timestamp
            (server-generated on every PATCH), this one is taken from the
            request body verbatim - the caller (cluster_manager) sets it
            before sending.
        cpu_history:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/JobHistorySample'
        memory_history:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/JobHistorySample'
    JobInstanceAppend:
      type: object
      description: |
        Body of `PUT /api/v1/jobs/{job_id}/{instance_id}`. Only the last
        element of `instance_list` is appended; earlier elements are ignored.
      properties:
        instance_list:
          type: array
          items:
            $ref: '#/components/schemas/JobInstance'
...